ref : 튜터님의 말씀 정리 refactor : ai 요청시 api요청 서비스와 프롬프트 서비스를 분리
